Extract menu
