+++ snake/makefile
@@ -3,8 +3,8 @@ CFLAGS = -std=c++17 -Wall -pedantic
 
 all: snake
 
-snake: set_ticker.o snake.o res.o game.o main.cpp graph.o
-	$(CC) -o snake -g set_ticker.o snake.o res.o game.o graph.o main.cpp $(CFLAGS) -lcurses
+snake: set_ticker.o snake.o res.o game.o main.cpp graph.o menu.o
+	$(CC) -o snake -g set_ticker.o snake.o res.o game.o graph.o menu.o main.cpp $(CFLAGS) -lcurses
 
 snake.o: snake.cpp snake.h
 	$(CC) -c -g snake.cpp $(CFLAGS)
@@ -20,6 +20,9 @@ game.o: game.cpp game.h
 
 graph.o: graph.h graph.cpp
 	$(CC) -c -g graph.cpp $(CFLAGS)
+
+menu.o: menu.cpp menu.h
+	$(CC) -c -g menu.cpp $(CFLAGS)
 
 .PHONY: clean
 clean:
